requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
click>=8.0.0
pyyaml>=6.0.0